// resources/game/utils/images.ts
type TImages = {
    background: string;
    slopingRoadRight: string;
    road: string;
    slopingRoadLeft: string;
    endRoadLeft: string;
    beforeSlopeRight: string;
    dirt: string;
    beforeSlopeLeft: string;
    endRoadRight: string;
    cliffRight: string;
    stoneRoadStart: string;
    stoneRoad: string;
    stoneRoadEnd: string;
    cliffLeft: string;
    groundFlatLeftSection: string;
    characterIdle: string;
    characterWalking: string;
    characterJumpStart: string;
    characterJumpLoop: string;
    decor02: string;
    decor03: string;
    decor04: string;
    maison01: string;
    maison02: string;

    hell_slopingRoadRight: string;
    hell_road: string;
    hell_slopingRoadLeft: string;
    hell_endRoadLeft: string;
    hell_beforeSlopeRight: string;
    hell_dirt: string;
    hell_beforeSlopeLeft: string;
    hell_endRoadRight: string;
    hell_cliffRight: string;
    hell_stoneRoadStart: string;
    hell_stoneRoad: string;
    hell_stoneRoadEnd: string;
    hell_cliffLeft: string;
    hell_decor01: string;
    hell_decor02: string;
    hell_decor03: string;
    hell_decor04: string;
};

export const images: TImages = {
    background: "/gameAssets/environment/Background.png",
    slopingRoadRight: "/gameAssets/groundElements/Ground_01.png",
    road: "/gameAssets/groundElements/Ground_02.png",
    slopingRoadLeft: "/gameAssets/groundElements/Ground_03.png",
    endRoadLeft: "/gameAssets/groundElements/Ground_04.png",
    beforeSlopeRight: "/gameAssets/groundElements/Ground_05.png",
    dirt: "/gameAssets/groundElements/Ground_06.png",
    beforeSlopeLeft: "/gameAssets/groundElements/Ground_07.png",
    endRoadRight: "/gameAssets/groundElements/Ground_08.png",
    cliffRight: "/gameAssets/groundElements/Ground_09.png",
    stoneRoadStart: "/gameAssets/groundElements/Ground_10.png",
    stoneRoad: "/gameAssets/groundElements/Ground_11.png",
    stoneRoadEnd: "/gameAssets/groundElements/Ground_12.png",
    cliffLeft: "/gameAssets/groundElements/Ground_13.png",
    groundFlatLeftSection: "/gameAssets/groundElements/GroundFlatLeftSection.png",
    characterIdle: "/gameAssets/character/idle.png",
    characterWalking: "/gameAssets/character/walking.png",
    characterJumpStart: "/gameAssets/character/jumpStart.png",
    characterJumpLoop: "/gameAssets/character/jumpLoop.png",
    decor02: "/gameAssets/environment/decor02.png",
    decor03: "/gameAssets/environment/decor03.png",
    decor04: "/gameAssets/environment/decor04.png",
    maison01: "/gameAssets/environment/maison01.png",
    maison02: "/gameAssets/environment/maison02.png",
    hell_slopingRoadRight: "/gameAssets/hell/groundElements/Ground_01.png",
    hell_road: "/gameAssets/hell/groundElements/Ground_02.png",
    hell_slopingRoadLeft: "/gameAssets/hell/groundElements/Ground_03.png",
    hell_endRoadLeft: "/gameAssets/hell/groundElements/Ground_04.png",
    hell_beforeSlopeRight: "/gameAssets/hell/groundElements/Ground_05.png",
    hell_dirt: "/gameAssets/hell/groundElements/Ground_06.png",
    hell_beforeSlopeLeft: "/gameAssets/hell/groundElements/Ground_07.png",
    hell_endRoadRight: "/gameAssets/hell/groundElements/Ground_08.png",
    hell_cliffRight: "/gameAssets/hell/groundElements/Ground_09.png",
    hell_stoneRoadStart: "/gameAssets/hell/groundElements/Ground_10.png",
    hell_stoneRoad: "/gameAssets/hell/groundElements/Ground_11.png",
    hell_stoneRoadEnd: "/gameAssets/hell/groundElements/Ground_12.png",
    hell_cliffLeft: "/gameAssets/hell/groundElements/Ground_13.png",
    hell_decor01: "/gameAssets/hell/environment/Sign_02.png",
    hell_decor02: "/gameAssets/hell/environment/Sign_04.png",
    hell_decor03: "/gameAssets/hell/environment/Statue.png",
    hell_decor04: "/gameAssets/hell/environment/Demonic_Tree_02.png",
};